export_test/bot_оо121211122112/requirements.txt
pyTelegramBotAPI>=4.0.0
requests>=2.25.0
python-dotenv>=0.19.0
